requirements.txt
async-timeout==4.0.2
bcrypt==4.0.1
beautifulsoup4==4.12.0
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dash==2.9.2
dash-core-components==2.0.0
dash-cytoscape==0.3.0
dash-html-components==2.0.0
dash-mantine-components==0.12.0
dash-table==5.0.0
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
maxdmg-resource==0.0.3
plotly==5.13.1
pyairtable==1.4.0
python-dotenv==1.0.0
redis==4.5.3
requests==2.28.2
soupsieve==2.4
SQLAlchemy==2.0.7
tenacity==8.2.2
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
WTForms==3.0.1
